Separated organization data updating from harvesting
