+++ ckanext/xroad_integration/harvesters/xroad_harvester.py
@@ -54,18 +54,6 @@ adapter = TimeoutHTTPAdapter(max_retries=retry_strategy)
 http = requests.Session()
 http.mount("http://", adapter)
 
-PUBLIC_ORGANIZATION_CLASSES = ['GOV', 'MUN', 'ORG']
-COMPANY_CLASSES = ['COM']
-
-def _convert_xroad_value_to_uniform_list(value):
-    if isinstance(value, basestring):
-        return []
-
-    if type(value) is dict:
-        return [value]
-
-    return value
-
 
 class XRoadHarvesterPlugin(HarvesterBase):
     config = None
@@ -161,123 +149,6 @@ class XRoadHarvesterPlugin(HarvesterBase):
                 'xroad_memberclass': member['memberClass'],
                 'xroad_membercode': member['memberCode']
             }
-
-            # Fetch organization information
-
-            if member['memberClass'] in PUBLIC_ORGANIZATION_CLASSES:
-                try:
-                    org_information_list = self._get_organization_information(harvest_job.source.url, member['memberCode'])
-
-                    organization_info = None
-
-                    if org_information_list:
-                        # If PTV has only one matching organization
-                        if type(org_information_list) is dict:
-                            log.info("Only one matching organization in ptv for organization %s" % member['name'])
-                            organization_info = org_information_list
-                        else:
-                            log.info("Multiple organizations found in ptv for organization %s" % member['name'] )
-                            # Match only if PTV title matches our organization title
-                            organization_info = self._parse_organization_info(org_information_list, member['name'])
-
-                        if organization_info:
-                            log.info("Parsing organization information for %s" % member['name'])
-
-                            if organization_info.get('organizationNames', {}):
-                                org_names = _convert_xroad_value_to_uniform_list(
-                                    organization_info.get('organizationNames', {}).get('organizationName', {}))
-
-                                org_names_translated = {
-                                    "fi": next((name.get('value', '') for name in org_names if (name.get('language') == 'fi') and name.get('type') == "Name"), ""),
-                                    "sv": next((name.get('value', '') for name in org_names if (name.get('language') == 'sv') and name.get('type') == "Name"), ""),
-                                    "en": next((name.get('value', '') for name in org_names if (name.get('language') == 'en') and name.get('type') == "Name"), ""),
-                                }
-
-                                organization_dict['title_translated'] = org_names_translated
-
-                            if organization_info.get('organizationDescriptions', {}):
-                                org_descriptions = _convert_xroad_value_to_uniform_list(
-                                    organization_info.get('organizationDescriptions', {}).get('organizationDescription', {}))
-
-                                org_descriptions_translated = {
-                                    "fi": next((description.get('value', '') for description in org_descriptions if description.get('language') == 'fi'), ""),
-                                    "sv": next((description.get('value', '') for description in org_descriptions if description.get('language') == 'sv'), ""),
-                                    "en": next((description.get('value', '') for description in org_descriptions if description.get('language') == 'en'), "")
-                                }
-
-                                organization_dict['description_translated'] = org_descriptions_translated
-
-                            if organization_info.get('webPages', {}):
-                                webpages = _convert_xroad_value_to_uniform_list(
-                                    organization_info.get('webPages', {}).get('webPage', {}))
-
-                                webpage_addresses = {
-                                    "fi": next((webpage.get('url', '') for webpage in webpages if webpage.get('language') == 'fi'), ""),
-                                    "sv": next((webpage.get('url', '') for webpage in webpages if webpage.get('language') == 'sv'), ""),
-                                    "en": next((webpage.get('url', '') for webpage in webpages if webpage.get('language') == 'en'), "")
-                                }
-
-                                webpage_descriptions = {
-                                    "fi": next((webpage.get('value', '') for webpage in webpages if webpage.get('language') == 'fi'), ""),
-                                    "sv": next((webpage.get('value', '') for webpage in webpages if webpage.get('language') == 'sv'), ""),
-                                    "en": next((webpage.get('value', '') for webpage in webpages if webpage.get('language') == 'en'), "")
-                                }
-
-
-                                organization_dict['webpage_address'] = webpage_addresses
-                                organization_dict['webpage_description'] = webpage_descriptions
-
-                            organization_dict['organization_guid'] = organization_info.get('guid', '')
-
-
-                except ContentFetchError:
-                    self._save_gather_error("Failed to fetch organization information with id %s" % member['membercode'], harvest_job)
-
-            elif member['memberClass'] in COMPANY_CLASSES:
-                try:
-                    company = self._get_companies_information(harvest_job.source.url, member['memberCode'])
-                    if type(company) is dict:
-                        if company.get('companyForms'):
-                            company_forms = _convert_xroad_value_to_uniform_list(company.get('companyForms', {}).get('companyForm', {}))
-                            forms = {
-                                "fi": next((form.get('name') for form in company_forms if form.get('language') == 'FI'), ""),
-                                "sv": next((form.get('name') for form in company_forms if form.get('language') == 'SE'), ""),
-                                "en": next((form.get('name') for form in company_forms if form.get('language') == 'EN'), "")
-                            }
-
-                            organization_dict['company_type'] = forms
-
-                        if company.get('businessAddresses'):
-                            business_addresses = _convert_xroad_value_to_uniform_list(company.get('businessAddresses', {}).get('businessAddress', {}))
-
-                            business_address = business_addresses[0] if business_addresses else None
-
-                            # TODO: language should be country
-                            organization_dict['postal_address'] = business_address.get('street') + ', ' + str(business_address.get('postCode'))\
-                                                                  + ', ' + business_address.get('city') + ', ' + business_address.get('language')
-
-                        if company.get('languages'):
-                            languages = _convert_xroad_value_to_uniform_list(company.get('languages', {}).get('language', {}))
-                            company_languages = {
-                                "fi": next((language.get('name', '') for language in languages if language.get('language') == 'FI'), ""),
-                                "sv": next((language.get('name', '') for language in languages if language.get('language') == 'SE'), ""),
-                                "en": next((language.get('name', '') for language in languages if language.get('language') == 'EN'), "")
-                            }
-
-                            organization_dict['company_language'] = company_languages
-
-                        # Convert "2001-06-11T00:00:00.000+03:00" to "2001-06-11T00:00:00"
-                        organization_dict['company_registration_date'] = company.get('registrationDate').split(".")[0]
-
-                        if company.get('businessIdChanges'):
-                            business_id_changes = _convert_xroad_value_to_uniform_list(
-                                company.get('businessIdChanges', {}).get('businessIdChange', {}))
-
-                            old_business_ids = [str(business_id_change.get('oldBusinessId')) for business_id_change in business_id_changes]
-                            organization_dict['old_business_ids'] = json.dumps(old_business_ids)
-
-                except ContentFetchError:
-                    self._save_gather_error("Failed to fetch company information with id %s" % member['membercode'], harvest_job)
 
 
             org = self._create_or_update_organization(organization_dict, harvest_job)
@@ -671,78 +542,6 @@ class XRoadHarvesterPlugin(HarvesterBase):
             raise ContentFetchError("Calling XRoad service IsProvider failed")
 
         return ""
-
-    @staticmethod
-    def _get_organization_information(url, business_code):
-        try:
-            r = http.get(url + '/Consumer/GetOrganizations', params = {'businessCode': business_code},
-            headers = {'Accept': 'application/json'})
-
-            response_json = r.json()
-            if response_json.get("error"):
-                log.info(response_json.get("error").get("string"))
-                return []
-
-            if response_json.get('organizationList', {}).get('organization') is dict:
-                return [response_json['organizationList']['organization']]
-
-            return response_json.get('organizationList', {}).get('organization')
-        except ConnectionError:
-            raise ContentFetchError("Calling XRoad service GetOrganizations failed")
-
-    @staticmethod
-    def _parse_organization_info(data, organization_name):
-        organization_info = next((org_info for org_info in data for name in
-                                  _convert_xroad_value_to_uniform_list(org_info.get('organizationNames', {})
-                                                                       .get('organizationName'))
-                                  if name.get('value', {}) == organization_name), None)
-
-        return  organization_info
-
-    @staticmethod
-    def _get_companies_information(url, business_id):
-        try:
-            r = http.get(url + '/Consumer/GetCompanies', params = {'businessId': business_id},
-                         headers = {'Accept': 'application/json'})
-
-            response_json = r.json()
-            if response_json.get("error"):
-                log.info(response_json.get("error").get("string"))
-                return ""
-
-            return response_json.get('companyList', {}).get('company')
-        except ConnectionError:
-            raise ContentFetchError("Calling XRoad service GetCompanies failed")
-
-    @staticmethod
-    def _get_organization_changes(url, guid, changed_after ):
-        try:
-            r = http.get(url + '/Consumer/HasOrganizationChanged', params = {'guid': guid, 'changedAfter': changed_after},
-                         headers = {'Accept': 'application/json'})
-
-            response_json = r.json()
-            if response_json.get("error"):
-                log.info(response_json.get("error").get("string"))
-                return ""
-
-            return r.json()
-        except ConnectionError:
-            raise ContentFetchError("Calling XRoad service HasOrganizationChanged failed")
-
-    @staticmethod
-    def _get_company_changes(url, business_id, changed_after ):
-        try:
-            r = http.get(url + '/Consumer/HasCompanyChanged', params = {'business_id': business_id, 'changedAfter': changed_after},
-                         headers = {'Accept': 'application/json'})
-
-            response_json = r.json()
-            if response_json.get("error"):
-                log.info(response_json.get("error").get("string"))
-                return ""
-
-            return r.json()
-        except ConnectionError:
-            raise ContentFetchError("Calling XRoad service HasCompanyChanged failed")
 
     @classmethod
     def _last_error_free_job(cls, harvest_job):
